+++ cinetica/__init__.py
@@ -1,5 +1,14 @@
-from .movimiento_rectilineo import MovimientoRectilineo
-from .movimiento_parabolico import MovimientoParabolico
-from .movimiento_circular import MovimientoCircular
+from .rectilineo import MovimientoRectilineoUniforme, MovimientoRectilineoUniformementeVariado
+from .parabolico import MovimientoParabolicoBase, MovimientoParabolicoAnalisis
+from .circular import MovimientoCircularUniforme, MovimientoCircularUniformementeVariado
 
-__version__ = "0.3.6"
+__version__ = "0.3.8"
+
+__all__ = [
+    "MovimientoRectilineoUniforme",
+    "MovimientoRectilineoUniformementeVariado",
+    "MovimientoParabolicoBase",
+    "MovimientoParabolicoAnalisis",
+    "MovimientoCircularUniforme",
+    "MovimientoCircularUniformementeVariado",
+]